requirements.txt
flask
mysql-connector-python
python-dateutil
